+++ packages/code-du-travail-frontend/__tests__/a-propos.test.tsx
@@ -1,5 +1,4 @@
 import { render } from "@testing-library/react";
-import React from "react";
 
 import About from "../pages/a-propos";
