--- packages/code-du-travail-frontend/__tests__/code-du-travail.test.js
@@ -1,11 +1,0 @@
-import { render } from "@testing-library/react";
-import React from "react";
-
-import CodeDuTravail from "../pages/code-du-travail/[slug]";
-
-describe("<CodeDuTravail />", () => {
-  it("should render", () => {
-    const { container } = render(<CodeDuTravail />);
-    expect(container).toMatchSnapshot();
-  });
-});
