--- packages/code-du-travail-frontend/__tests__/convention-collective.test.js
@@ -1,11 +1,0 @@
-import { render } from "@testing-library/react";
-import React from "react";
-
-import ConventionCollective from "../pages/convention-collective/[slug]";
-
-describe("<ConventionCollective />", () => {
-  it("should render", () => {
-    const { container } = render(<ConventionCollective />);
-    expect(container).toMatchSnapshot();
-  });
-});
